Simplified configuration folder
