+++ config/dev.config.js
@@ -2,8 +2,8 @@ import FaviconsWebpackPlugin from 'favicons-webpack-plugin';
 import ExtendedDefinePlugin from 'extended-define-webpack-plugin';
 import PrerenderSpaPlugin from 'prerender-spa-plugin';
 
-import { environments } from '../';
-import { getVariables } from '../variables';
+import { environments } from './environments';
+import { getVariables } from './runtime';
 import { app, favicon } from './settings';
 import { baseConfig } from './base.config';
