+++ config/index.js
@@ -1,0 +1,8 @@
+import { environments } from './environments';
+
+export const webpackInit = (requestedEnv) => {
+  process.env.NODE_ENV = environments[requestedEnv];
+
+  // eslint-disable-next-line global-require,import/no-dynamic-require
+  return require(`./${requestedEnv}.config`).config;
+};
