+++ config/local.config.js
@@ -2,8 +2,8 @@ import ExtendedDefinePlugin from 'extended-define-webpack-plugin';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import LiveReloadPlugin from 'webpack-livereload-plugin';
 
-import { environments } from '../';
-import { getVariables } from '../variables';
+import { environments } from './environments';
+import { getVariables } from './runtime';
 import { baseConfig } from './base.config';
 
 export const config = {
