+++ config/settings/index.js
@@ -2,4 +2,6 @@ export * from './app';
 export * from './favicon';
 export * from './manifest';
 export * from './page';
+export * from './paths';
+export * from './project';
 export * from './rules';
