--- environments/config/index.js
@@ -1,3 +1,0 @@
-export * from './page';
-export * from './paths';
-export * from './project';
