--- environments/webpack/settings/rules.js
@@ -1,34 +1,0 @@
-
-const styleLoaders = minify => ([
-  { loader: 'style-loader' },
-  {
-    loader: 'css-loader',
-    options: {
-      modules: true,
-      importLoaders: 1,
-      minimize: minify,
-      localIdentName: minify ? '[hash:base64:24]' : '[path][name]-[local]',
-    },
-  },
-  { loader: 'sass-loader' },
-]);
-
-const getRules = minify => ([
-  {
-    test: /\.js$/,
-    use: 'babel-loader',
-  },
-  {
-    test: /\.scss$/,
-    use: styleLoaders(minify),
-  },
-  {
-    test: /\.(json|png|jpg|gif|svg|woff|woff2)$/,
-    use: 'file-loader',
-  },
-]);
-
-export const rules = {
-  pretty: getRules(false),
-  minified: getRules(true),
-};
